brain/requirements.txt
streamlit
pandas-ta
ccxt
sqlalchemy
requests
plotly
pyyaml
